requirements.txt
caslib.py==2.2.2
oauth2client
djangorestframework
jwt.py
python-ldap
requests
threepio
apache-libcloud==0.20.1

# For developers -- Uncomment
# twine
# wheel
